Refactor imports in client and index files for improved structure and clarity
